+++ website/site-data.js
@@ -5,8 +5,8 @@ globalThis.OSTRIN_SITE_FACTS = Object.freeze({
   "releaseDate": "2026-09-24",
   "releaseUrl": "https://github.com/sircalch/Ostrin/releases/tag/v0.1.0",
   "designDocs": "22",
-  "examples": "209",
-  "integrationTests": "201",
+  "examples": "211",
+  "integrationTests": "202",
   "differentialTests": "6",
   "unitTests": "2"
 });